# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/filters/{filterName}/flag-instances": {
      "post": {
        "summary": "Read Flagged Instances",
        "description": "Flags instances using a specified filter and options.\n\nArgs:\n    filter_name: The name of the filter to use.\n    options: The options for flagging instances, which will be validated against the filter's schema.\n    db: Database session for any necessary database access during filtering.\n    current_user: The user making the request, which may be relevant for certain filters.",
        "operationId": "read_flagged_instances_filters__filterName__flag_instances_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "filterName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filtername"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Options"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {},
                  "title": "Response Read Flagged Instances Filters  Filtername  Flag Instances Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/filters/{filterName}/get-contexts": {
      "post": {
        "summary": "Read Contexts",
        "description": "Retrieves contexts for a list of instances using a specified filter and options.\n\nArgs:\n    filter_name: The name of the filter to use.\n    instances: The list of instances to retrieve contexts for. Each instance will be validated against the filter's instance schema.\n    options: The options for retrieving contexts, which will be validated against the filter's schema.\n    db: Database session for any necessary database access during context retrieval.\n    current_user: The user making the request, which may be relevant for certain filters.",
        "operationId": "read_contexts_filters__filterName__get_contexts_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "filterName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filtername"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceOptions"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {},
                  "title": "Response Read Contexts Filters  Filtername  Get Contexts Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/filters/{filterName}/decide-instances": {
      "post": {
        "summary": "Read Decisions",
        "description": "Decides whether instances pass a filter in given contexts using specified options.\n\nArgs:\n    filter_name: The name of the filter to use.\n    body: An InstanceContextOptions object containing the list of instance-context pairs and the options for deciding instances. Each instance and context will be validated against the filter's schemas.\n    db: Database session for any necessary database access during the decision process.\n    current_user: The user making the request, which may be relevant for certain filters.",
        "operationId": "read_decisions_filters__filterName__decide_instances_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "filterName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filtername"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceContextOptions"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "boolean"
                  },
                  "title": "Response Read Decisions Filters  Filtername  Decide Instances Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/filters/{filterName}/apply": {
      "post": {
        "summary": "Apply Filter",
        "description": "Applies a filter to a label group for a list of instances using specified options.\n\nArgs:\n    filter_name: The name of the filter to apply.\n    body: An InstanceOptions object containing the list of instances and the options for applying the filter. Each instance will be validated against the filter's instance schema, and the options will be validated against the filter's apply filter options schema.\n    db: Database session for any necessary database access during filter application.\n    current_user: The user making the request, which may be relevant for certain filters.",
        "operationId": "apply_filter_filters__filterName__apply_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "filterName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filtername"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceOptions"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/edit-chapter-data/{chapterId}": {
      "get": {
        "summary": "Read Edit Chapter Data",
        "description": "Gets all data associated with a chapter required for editing.\n\nArgs:\n    chapter_id: Chapter id of requested chapter\n    novel_id: Novel id of requested chapter\n\nRaises:\n    404: Chapter not found (or insufficient permissions).",
        "operationId": "read_edit_chapter_data_edit_chapter_data__chapterId__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "chapterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chapterid"
            }
          },
          {
            "name": "novelId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Novelid"
            }
          },
          {
            "name": "labelGroupsNum",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "title": "Labelgroupsnum"
            }
          },
          {
            "name": "subjectId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subjectid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditChapterData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Chapter": {
        "properties": {
          "chapterId": {
            "type": "string",
            "format": "uuid",
            "title": "Chapterid"
          },
          "chapterNum": {
            "type": "integer",
            "title": "Chapternum"
          },
          "chapterTitle": {
            "type": "string",
            "title": "Chaptertitle"
          },
          "chapterIsPublic": {
            "type": "boolean",
            "title": "Chapterispublic"
          },
          "novelId": {
            "type": "string",
            "format": "uuid",
            "title": "Novelid"
          }
        },
        "type": "object",
        "required": [
          "chapterId",
          "chapterNum",
          "chapterTitle",
          "chapterIsPublic",
          "novelId"
        ],
        "title": "Chapter",
        "description": "Pydantic schema for chapter metadata. Represents a single \"chapter\" entry, which groups all its revisions.\n\nAttributes:\n    chapter_id: UUID primary key identifier.\n    chapter_num: The chapter number.\n    novel_id: UUID foreign key to the novel this chapter belongs to."
      },
      "ChapterContent": {
        "properties": {
          "chapterContentText": {
            "type": "string",
            "title": "Chaptercontenttext"
          },
          "chapterContentVersion": {
            "type": "integer",
            "title": "Chaptercontentversion"
          },
          "chapterContentId": {
            "type": "string",
            "format": "uuid",
            "title": "Chaptercontentid"
          }
        },
        "type": "object",
        "required": [
          "chapterContentText",
          "chapterContentVersion",
          "chapterContentId"
        ],
        "title": "ChapterContent",
        "description": "Pydantic schema for the text content of a chapter.\n\nAttributes:\n    chapter_content_text: The full text content of the chapter.\n    chapter_content_version: The version number of the text content, used for optimistic concurrency control when updating text.\n    chapter_content_id: The UUID of the text content, used for optimistic concurrency control when updating text."
      },
...
      "EditChapterData": {
        "properties": {
          "chapter": {
            "$ref": "#/components/schemas/Chapter"
          },
          "chapterContent": {
            "$ref": "#/components/schemas/ChapterContent"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "labelGroupList": {
            "items": {
              "$ref": "#/components/schemas/LabelGroupListEntry"
            },
            "type": "array",
            "title": "Labelgrouplist"
          },
          "labelDataList": {
            "items": {
              "$ref": "#/components/schemas/LabelDataEntry"
            },
            "type": "array",
            "title": "Labeldatalist"
          }
        },
        "type": "object",
        "required": [
          "chapter",
          "chapterContent",
          "role",
          "labelGroupList",
          "labelDataList"
        ],
        "title": "EditChapterData",
        "description": "Pydantic schema for the data needed to edit a chapter.\n\nAttributes:\n    chapter: Chapter being edited.\n    chapter_content: ChapterContent being edited.\n    label_groups: List of LabelGroups in this novel.\n    labels: List of Labels in this chapter content."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InstanceContextOptions": {
        "properties": {
          "instanceContexts": {
            "items": {
              "prefixItems": [
                {},
                {}
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Instancecontexts"
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "instanceContexts",
          "options"
        ],
        "title": "InstanceContextOptions"
      },
      "InstanceOptions": {
        "properties": {
          "instances": {
            "items": {},
            "type": "array",
            "title": "Instances"
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "instances",
          "options"
        ],
        "title": "InstanceOptions"
      },
      "Label": {
        "properties": {
          "labelEntityGroup": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Labelentitygroup"
          },
          "labelScore": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Labelscore"
          },
          "labelWord": {
            "type": "string",
            "maxLength": 128,
            "title": "Labelword"
          },
          "labelStart": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Labelstart"
          },
          "labelEnd": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Labelend"
          },
          "labelDirty": {
            "type": "boolean",
            "title": "Labeldirty"
          },
          "labelDataId": {
            "type": "string",
            "format": "uuid",
            "title": "Labeldataid"
          },
          "labelId": {
            "type": "string",
            "format": "uuid",
            "title": "Labelid"
          }
        },
        "type": "object",
        "required": [
          "labelEntityGroup",
          "labelScore",
          "labelWord",
          "labelStart",
          "labelEnd",
          "labelDirty",
          "labelDataId",
          "labelId"
        ],
        "title": "Label",
        "description": "Pydantic schema for a label that belongs to a LabelData.\nExtends LabelBase with a label_data_id foreign key and label_id pkey."
      },
...
      "LabelData": {
        "properties": {
          "labelDataId": {
            "type": "string",
            "format": "uuid",
            "title": "Labeldataid"
          },
          "labelGroupId": {
            "type": "string",
            "format": "uuid",
            "title": "Labelgroupid"
          },
          "chapterContentId": {
            "type": "string",
            "format": "uuid",
            "title": "Chaptercontentid"
          }
        },
        "type": "object",
        "required": [
          "labelDataId",
          "labelGroupId",
          "chapterContentId"
        ],
        "title": "LabelData",
        "description": "Pydantic schema for a list of labels in some text.\n\nAttributes:\n    label_data_id: UUID identifier for this LabelData.\n    label_group_id: UUID of label group this LabelData belongs to.\n    chapter_content_id: UUID of chapter content this LabelData is labelling."
      },
      "LabelDataEntry": {
        "properties": {
          "labelDataId": {
            "type": "string",
            "format": "uuid",
            "title": "Labeldataid"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/Label"
            },
            "type": "array",
            "title": "Labels"
          }
        },
        "type": "object",
        "required": [
          "labelDataId",
          "labels"
        ],
        "title": "LabelDataEntry"
      },
      "LabelGroup": {
        "properties": {
          "labelGroupId": {
            "type": "string",
            "format": "uuid",
            "title": "Labelgroupid"
          },
          "labelGroupName": {
            "type": "string",
            "maxLength": 31,
            "title": "Labelgroupname"
          },
          "novelId": {
            "type": "string",
            "format": "uuid",
            "title": "Novelid"
          }
        },
        "type": "object",
        "required": [
          "labelGroupId",
          "labelGroupName",
          "novelId"
        ],
        "title": "LabelGroup",
        "description": "Pydantic schema for a label group.\n\nAttributes:\n    label_group_id: UUID identifier for this label group.\n    label_group_name: Name of this label group.\n    novel_id: UUID of novel this label group belongs to."
      },
      "LabelGroupListEntry": {
        "properties": {
          "labelGroup": {
            "$ref": "#/components/schemas/LabelGroup"
          },
          "labelData": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LabelData"
              },
              {
                "type": "null"
              }
            ]
          },
          "role": {
            "$ref": "#/components/schemas/LabelRole"
          }
        },
        "type": "object",
        "required": [
          "labelGroup",
          "labelData",
          "role"
        ],
        "title": "LabelGroupListEntry"
      },
      "LabelRole": {
        "type": "string",
        "enum": [
          "owner",
          "viewer",
          "editor"
        ],
        "title": "LabelRole"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "owner",
          "viewer",
          "editor"
        ],
        "title": "Role"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}